% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{compare_lc_tibble}
\alias{compare_lc_tibble}
\title{compare_lc_tibble}
\usage{
compare_lc_tibble(experimental_lc, standard_lc)
}
\arguments{
\item{experimental_lc}{experimental_lc.}

\item{standard_lc}{standard_lc.}
}
\value{
A list.
}
\description{
The minimum and maximum lc of two lc_tibble should be kept consistent when comparing them.
}
\examples{
data("experimental_data")
data("experimental_car")
data("standard_car")
ResList <- lc2rt(lc_tibble = experimental_car, targetRt = max(experimental_data$rt))
experimental_car <- ResList$lc_tibble
ResList$p
ResList <- lc2rt(lc_tibble = standard_car, targetRt = max(experimental_data$rt))
standard_car <- ResList$lc_tibble
ResList$p
experimental_car <- deadTime_add(lc_tibble = experimental_car, deadTime = min(experimental_data$rt))
standard_car <- deadTime_add(lc_tibble = standard_car, deadTime = min(experimental_data$rt))
ResList <- compare_lc_tibble(experimental_lc = experimental_car, standard_lc = standard_car)
experimental_car <- ResList$experimental_lc
standard_car <- ResList$standard_lc
}
