% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{lc2rt}
\alias{lc2rt}
\title{lc2rt}
\usage{
lc2rt(lc_tibble, targetRt)
}
\arguments{
\item{lc_tibble}{A lc_tibble.}

\item{targetRt}{The maximum retention time user want to overwrite.}
}
\value{
A list.
}
\description{
Use lc_tibble to predict rt.
}
\examples{
data("experimental_data")
data("experimental_car")
data("standard_car")
ResList <- lc2rt(lc_tibble = experimental_car, targetRt = max(experimental_data$rt))
experimental_car <- ResList$lc_tibble
ResList$p
}
