% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{orignRT2newRT}
\alias{orignRT2newRT}
\title{orignRT2newRT}
\usage{
orignRT2newRT(experimental_data, experimental_lc, standard_lc, thread = 1)
}
\arguments{
\item{experimental_data}{experimental_data tibble.}

\item{experimental_lc}{This is the lc_tibble of the correction substance corresponding to the experimental data.}

\item{standard_lc}{This is the lc_tibble of the correction substance corresponding to the standard chromatographic condition.}

\item{thread}{The number of parallel thread.}
}
\value{
A new experimental data.
}
\description{
Convert experimental retention time to standard retention time.
}
\examples{
data("experimental_data")
data("experimental_car")
data("standard_car")
ResList <- lc2rt(lc_tibble = experimental_car, targetRt = max(experimental_data$rt))
experimental_car <- ResList$lc_tibble
ResList$p
ResList <- lc2rt(lc_tibble = standard_car, targetRt = max(experimental_data$rt))
standard_car <- ResList$lc_tibble
ResList$p
experimental_car <- deadTime_add(lc_tibble = experimental_car, deadTime = min(experimental_data$rt))
standard_car <- deadTime_add(lc_tibble = standard_car, deadTime = min(experimental_data$rt))
ResList <- compare_lc_tibble(experimental_lc = experimental_car, standard_lc = standard_car)
experimental_car <- ResList$experimental_lc
standard_car <- ResList$standard_lc
experimental_data$ri <- sapply(experimental_data$rt, function(x) {
  rt2ri(x, experimental_lc = experimental_car)
})
experimental_data <- orignRT2newRT(experimental_data = experimental_data,
                                   experimental_lc = experimental_car,
                                   standard_lc = standard_car, thread = 1)
}
