% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoRI.R
\name{orignRT2newRT_old}
\alias{orignRT2newRT_old}
\title{Title}
\usage{
orignRT2newRT_old(
  experimental_data,
  experimental_car,
  standard_car,
  thread = 1
)
}
\arguments{
\item{experimental_data}{test.}

\item{experimental_car}{test.}

\item{standard_car}{test.}

\item{thread}{test.}
}
\value{
test.
}
\description{
Title
}
\examples{
"A test"
}
