% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{rt2ri}
\alias{rt2ri}
\title{Title}
\usage{
rt2ri(rt, experimental_lc)
}
\arguments{
\item{rt}{A vector of experimental data rt.}

\item{experimental_lc}{This is the lc_tibble of the correction metabolite corresponding to the experimental data.}
}
\value{
A vector of ri.
}
\description{
Title
}
\examples{
data("experimental_data")
data("experimental_car")
data("standard_car")
ResList <- lc2rt(lc_tibble = experimental_car, targetRt = max(experimental_data$rt))
experimental_car <- ResList$lc_tibble
ResList$p
ResList <- lc2rt(lc_tibble = standard_car, targetRt = max(experimental_data$rt))
standard_car <- ResList$lc_tibble
ResList$p
experimental_car <- deadTime_add(lc_tibble = experimental_car, deadTime = min(experimental_data$rt))
standard_car <- deadTime_add(lc_tibble = standard_car, deadTime = min(experimental_data$rt))
ResList <- compare_lc_tibble(experimental_lc = experimental_car, standard_lc = standard_car)
experimental_car <- ResList$experimental_lc
standard_car <- ResList$standard_lc
experimental_data$ri <- sapply(experimental_data$rt, function(x) {
  rt2ri(x, experimental_lc = experimental_car)
})
}
